vu_project1_microservice/api_gateway/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
